Replace trainer.py with native PyTorch training system
Major changes:
- Remove trainer.py and implement train.py with native PyTorch training loop
- Convert ModelOutput from class to TypedDict for consistency with reference projects
- Add utility modules (train_utility.py, pytorch_utility.py) from reference projects
- Update config.py from iteration-based to epoch-based training
- Add torch_optimizer dependency for advanced optimizers (RAdam, Ranger)
- Implement torch.jit.script predictor optimization
- Add evaluator.judge property for model selection
- Fix type consistency issues and None handling for validation datasets
- Update target data type from float to long for cross_entropy compatibility

+++ src/hiho_pytorch_base/dataset.py
@@ -39,7 +39,7 @@ def preprocess(d: DatasetInput) -> DatasetOutput:
     """前処理関数"""
     output_data = DatasetOutput(
         feature=torch.from_numpy(d.feature).float(),
-        target=torch.from_numpy(d.target).float(),
+        target=torch.from_numpy(d.target).long(),
     )
     return output_data
